+++ backend/src/services/publications.service.ts
@@ -1,7 +1,12 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Publication } from '../models/publications/publication.schema';
 import { Model } from 'mongoose';
+import { UsersService } from './users.service';
 
 @Injectable()
 export class PublicationsService {
@@ -10,20 +15,27 @@ export class PublicationsService {
   constructor(
     @InjectModel('Publication')
     private readonly publicationModel: Model<Publication>,
+    private readonly userService: UsersService,
   ) {}
 
   async findAll() {
-    const publications = await this.publicationModel.find().exec();
+    const publications = await this.publicationModel
+      .find()
+      .sort({ date: -1 })
+      .exec();
+    return publications;
+  }
+
+  async findAllPubliUser(userId: string): Promise<Publication[]> {
+    const publications: any = await this.publicationModel
+      .find({ user: userId })
+      .exec();
     return publications;
   }
 
   async findOne(id: string): Promise<Publication> {
     let publication: Publication;
-    try {
-      publication = await this.publicationModel.findById(id);
-    } catch (error) {
-      throw new NotFoundException('Could not find publication');
-    }
+    publication = await this.publicationModel.findById(id);
     if (!publication) {
       throw new NotFoundException('Could not find publication');
     }
@@ -31,17 +43,76 @@ export class PublicationsService {
   }
 
   async insertPubli(title: string, text: string, user: string) {
-    const newPublication: Publication = new this.publicationModel({
+    const newPublication = new this.publicationModel({
       title: title,
       text: text,
       date: Date.now(),
       user: user,
     });
-    this.publications.push(newPublication);
     await newPublication.save();
   }
 
-  async deleteOnePubli(id: string) {
-    await this.publicationModel.deleteOne({ _id: id }).exec();
+  async deleteOnePubli(id: string, userId: string) {
+    const publication = await this.publicationModel
+      .find({ _id: id, user: userId })
+      .exec();
+    if (!publication) {
+      throw new UnauthorizedException('Unauthorized');
+    } else {
+      await this.publicationModel.deleteOne({ _id: id }).exec();
+      return true;
+    }
+  }
+
+  async countUserPubli(userId: string) {
+    const nbPubli = await this.publicationModel
+      .countDocuments({
+        user: userId,
+      })
+      .exec();
+    return nbPubli;
+  }
+
+  async countUserPubliLastWeek(userId: string) {
+    const nbPubli = await this.publicationModel
+      .countDocuments({
+        user: userId,
+        $expr: {
+          $gt: [
+            '$date',
+            {
+              $dateSubtract: {
+                startDate: '$$NOW',
+                unit: 'day',
+                amount: 7,
+              },
+            },
+          ],
+        },
+      })
+      .exec();
+    return nbPubli;
+  }
+
+  async findDataPieChart() {
+    var pipeline = [
+      {
+        $group: {
+          _id: '$user',
+          total_posts: { $sum: 1 },
+        },
+      },
+    ];
+    const usersPubli: { _id: string; total_posts: number }[] =
+      await this.publicationModel.aggregate(pipeline).exec();
+    let data: (string | number)[][] = [];
+    data.push(['User', 'Number of posts published']);
+    await Promise.all(
+      usersPubli.map(async (userPubli) => {
+        const username = await this.userService.getUsername(userPubli._id);
+        data.push([username, userPubli.total_posts]);
+      }),
+    );
+    return data;
   }
 }
